// keycloak-workstream/shared-services-assets/lib/interface.types.ts
export interface ConfigurationDocument {
  environment: Environment;
  vpc?: Vpc;
  helm?: Helm;
  docker?: Docker;
}
// ...
export interface Vpc {
  id: string;
  isolated?: boolean;
  subnets: {
    id: string;
  }[];
}

export interface Image {
  repository: string;
  tag: string;
}

export interface Docker {
  images: Image[];
}

export interface Helm {
  charts: HelmChart[];
}

export interface HelmChart {
  repositoryUrl: string;
  chartName: string;
  version: string;
  images: Image[];
}